Copy RTF To Clipboard fails - Second Commit
I have found another occurrence of this bug happening when you try to copy as html, followed by trying to copy as RTF. For whatever reason, the RTF Without Objects format throws an exception. Unless there is more insight onto the problem here, it appears removing it fixes the issue.